[FIX] undelete: Get correct token
The original implementation cdc2e9f4 used `undelete` token and added it as an
old token type. So on new implementations using the CSRF token this worked as
it mapped the token to CSRF. But old versions of MediaWiki which don't use the
CSRF token yet it'll fail because that token never existed.

The change in d4e6e60c broke then this also on newer versions as it only mapped
old token names to the CSRF token if they were actually old token name. And
that is not the case for `undelete` so it wasn't mapped to CSRF.

With 730bad3f the tokens for `delete` and `undelete` are the same and so it can
just use the `delete` token which will be mapped to CSRF on recent versions.

Bug: T105635
Change-Id: I9419aac0bc590fd3a3ed484012608f25dcf01cea