i18n: always follow master
We want to keep i18n up to date for all branches, so we should
make the i18n branch always follow master. This is consistent
with SVN's behavior. If the format of messages changes (e.g.
from %s to %(name)s), a /new/ key should be added, as otherwise
translations might break for older branches.

Change-Id: I5b26cda3f740f9421d70834bbfdfed53a6985933