Fix interwikiFormat support for Link
interwikiFormat is supposed to support a dictionary of Page or Link
objects.  There were several merges of core and compat 'textlib' to
keep the two in sync, such as core's 8978bfb6, and the support for
Link was broken somewhere along the way.

Add tests for Link in addition to existing tests for Page.

Also add a ValueError exception if it contains any other type.

Conflicts:
	pywikibot/textlib.py
	tests/textlib_tests.py

Change-Id: Ibb39ea8b3079083feec93249eb6abe4b997feb74
(manually cherry picked from 64ffa4279fb1)