[FIX] Page: Use repr-like if it can't be encoded
When the title contains characters which can't be encoded it falls back to use
an encoding which works like repr() on a unicode. It is also changing how it
behaves on Python 3 as it doesn't encode the title anymore and is thus not a
bytes anymore.

Conflicts:
	pywikibot/page.py
	tests/page_tests.py

Bug: T107428
Change-Id: I25dddac881891291c5a0dbe3f5dd2b1a0beedf0f
(cherry picked from commit 38589d3057847b9650b514e186b4bb3c66fad841)