pywikibot: Store ImportError in imported variable
If a lazy ImportError is to be handled, the ImportError
needs to be stored in a variable and then used when required.
Earlier, the imported variable was set to `None` and the
import error which was caught with `except` was used. But
this import error may not exist at a later time as python
guarantees it's existence only inside the `except` statement.
Hence, this is changed so that the ImportError is saved in the
variable that was being imported and checked later using
`isinstance(<variable>, ImportError)`.

(Re-enable F821 disabled in previous commit on 2.0 branch)

Bug: T134336
Change-Id: I80f82e7f0674bfca70688f28640620dacac1d80b
(manually cherry picked from commit 65cabc8258aa807a754a62182ae51fbf7df45ddd)